Simple Kokkos Team example.
This test does not check anything.  For now
it is just for testing compiling and running
on a GPU.

The test also shows that at the team level
the code runs on the GPU.  The team threads
then spawn member threads that also run
on the GPU.  This can be demonstrated by
trying to use std::cout instead of printf
and getting a compile error.