Simple unit test of virtual functions on device.
Create a unit test of a simple base class and
derived class with a virtual function.  This
class is then instantiated on device and the
pointer returned to host. The pointer is then
passed to a reduce on device where the virtual
function is called.  The result is checked that
the virtual and not base was called.