Simple unit test of virtual classes with views on device.
Now the virtual class is constructed with a std::vector
which is copied to a Kokkos::view on device.  This whole
class is then copied to the device using a copy constructor.
This device pointer is then passed to a kernel launch where
the virtual function is used to sum all values of the
view in parallel.