WIP: (do not merge)  Running a PrognosticProcess::run() on device.
Trying to determine what it would take to change the
PrognosticProcess, Prognostics, Atmosphere, Diagnostics,
Tendencies, and Model classes in order to execute
the virtual run() function and be able to access all
of the data in all of these classes.

The PrognosticProcess class and the Prognostics
class is mostly updated for use on a GPU without
breaking the CPU interface.  There is a question
of if this is the correct approach and this branch
is for exploring the options.

Currently there is a unit test that executes a
PrognosticProcess::run() on device and prints
some values from the Prognostics::interstitial_aerosols()
that is passed to the run() function.

The test runs but there are many compiler warnings about
calling __host__ from device since there are still many
std::vector and other data in these classes that will
need to be changed to views.