Fixes to compile and run on GPU.
There were many cases where functions were not marked KOKKOS_FUNCTION
and a few that were marked that should not have been.  KOKKOS_FUNCTION
and KOKKOS_INLINE_FUNCTION contain the __device__ macro that triggers
the compile of the function using the Cuda compiler.  This is necessary
to call the function on device but restricts what the function can
do so can not be used everywhere.

There is still a problem with the Diagnostics class.  The unit test does
not copy the whole class to device and I think it would fail if it
did. I think I know of a fix, but want to merge the major part of these
GPU changes into main before adding minor fixes.