Add ternary_nuc_merik2007 function to MAMNucleationProcess class.
This is a simple function used as part of the nucleation
process.  Mainly just a bunch of math that comes from a higher
order curve fitting of a multi-dimensional surface. We just
want to duplicate the F90 function in C++.  There is a test
that sweeps the parameter space with random numbers and compares
the results between the two implementations.