Rearrange to allow for both F90 vs. C++ and GPU vs. CPU tests.
On a platform that supports Fortran90 need to be able to
compile a test that compares the Fortran90 version of
this test to the C++ version.

On a platform that supports GPUs need to be able to
compile a test that compares the GPU version of this
test to the CPU version.

This is transitive.  Running both versions is the same
as comparing the Fortran90 version to the GPU version.