Fixed a memory error in the batch chemistry test.
The kinetic model data constructed in this test was being destroyed, leading
to invalid references to model data, so I added it to the list of members
maintained by the chemSolver class. This seems to have fixed the invalid
reads mentioned in #115, and may in fact fix that issue.

I also removed chemistry.hpp and chemistry.cpp, since they precede our current
chemistry approach and will be supplanted by it.