Initial implementation of C++ version of MAM Nucleation F90 code.
This ternary_nuc_merik2007 Fortran function from the MAM Nucleation
code was ported to a new MAMNucleationProcess class that is derived
from PrognosticProcess.  Eventually the rest of the Nucleation Fortran
code will be ported over to this class.

There is also a unit test comparing the results of calls to the Fortran
version with the C++ version.  And there is a unit test comparing the
CPU C++ version with a GPU compiled function.  The GPU test shows that
when compiled optimized there is a small numerical difference between
the GPU and CPU versions.  When the GPU code is compiled without
optimization, there are no differences.  There is a unit test showing
this optimization difference also.