Fix bugs found during Intel build
The AerosolSpecies had a serious bug
where the copy constructor and assignment
operator did not copy the whole class.
This was a bug introduced when string
was replaced by char[] and the default
constructor and assignment had to be
replaced. For some reason the regression
tests passed for the GCC compile.

Also needed to increase the tolerances
a bit to get to pass for an Intel build.
Note: using Intel-19.0.5