Finish adding PackType to MAM Nucleation Process class.
The MAM Nucleation Process class is the little test bed
class for demonstration purposes. It shows how SIMD
and GPUs might be used to execute processes once the
code is re-written from Fortran.

There is several levels of testing:
  1. Compare the original Fortran implementation
     with a Pack SIMD implementation but with
     a Pack size of 1, essentially making it
     just a double implementation like the
     Fortran.
  2. Compare the Pack SIMD length 1 implementation
     no CPU and GPU.  This shows that the GPU and
     CPU get almost the same results but there
     are tolerances in the floating point comparisons
     that are larger than comparing Fortran and C++.
  3. Compare Pack SIMD length 1 implementation to
     the PackSize compiled implementation. Typically
     a PackSize of 4 is compared to a PackSize of 1.
     This shows that the packing does not change
     the answer.  In this case, exact checks are
     made for bitwise matching.

In order to make the testing easier the functions
are templated on Pack which could be both for
a pack size of 4 and a pack size of 1 in a single
test so the answers could be compared.  Note that
it could have been templated on just the pack size
since the Pack class itself is required in the
functions.