Implemented a mechanism for setting parameters in aerosol processes.
This approach adds 3 methods to the aerosol process interface, all of the form
`set_param(name, value)`, where `name` is the name of a parameter that governs
the behavior of an aerosol process, and `value` is its value, which can be an
integer, a boolean flag, or a "real" (floating point) number.

In C++, one overrides the various forms of the `set_param` method to allow a
named parameter to be set by e.g. a host model. In Fortran, one implements
the three subroutines `set_integer_param`, `set_logical_param`, and
`set_real_param` in the module containing the aerosol process of interest. Code
is generated to implement C++ `set_param` overrides for these subroutines.

In addition, I've also renamed some parameters within the MAM4 nucleation Fortran
module to render them more sensible to humans, and I've gotten rid of some old
unused code from when we were considering diagnostic processes.