Fix mam_nucleation_process_tests for GPU.
Note:
(1) Changed the calling signature for the process run() function. Before
the prognostic class was non-const since prognostics are updated
during a process.  But Kokkos lambdas capture as const so the
class needs to be passed by const.  This is OK since everything in
the prognostic class that is updated is a Kokkos View and Kokkos
allows const views to be updated.

(2) Extended the custom deleter used in the copy_to_device() shared
pointer to call the destructor on the base class before calling
kokkos to free the allocated memory.