Add files for testing of Fortran version of Gas Aerosol Exchange process.
This adds the C++/Fortran interface files that will allow the testing
of the low level Fortran functions.  These are tested directly before
being tested by being called through the init and run routines.

Eventually once the Fortran is tested, the C++ versions will be created
and compared to the Fortran results.