Fix indexing problem.
The loop bounds were incorrectly using the index into the first
dimension of an array for the size() function instead of using 1.
Caused random seg faults.  Can do better anyway since the
code keeps track of the number of good entries in the array, it
does not need to check up to the maximum length.
e