Example of calling a process over all columns in a global model.
The question came up of how an application program might organize
all of the data needed for a global model in order to call
the Haero library for each column.  This is an example of how
that might be done.  It CPU compiles, but has not been GPU checked.

The idea is to create Kokkos::Views of the Haero classes that
can then be deep-copied to device which can then be indexed for
every column in the Kokkos league loop, calling the processes on
each column in the global model.  The league is the global model,
the team is one column and team members are levels in the column.

Some classes take Kokkos views in the constructor.  In this case
a global Kokkos view is created and a subview is taken for each
column.

Other classes allocate their own kokkos views as member data.
Care must be taken to init these classes consistently across
all of the model.  On the other hand this might be considered
a feature if columns over land are treated differently from columns
over the ocean.