Fixes for GPU compile warnings/errors.
Fixes:
"set_param_" partially overridden:  Since the base class
has multiple versions of this with different function signature
and the derived class only overrides a couple of them, the
compiler emits a warning.  The fix is to add a
   using AerosolProcess::set_param_;
line to quiet the warning.

Use of Kokkos::parallel_for in a KOKKOS_FUNCTION:
The run_ function is to be called from within
a Kokkos::parallel_for so can not contain a plain
parallel_for.  But the run_ gets a Team object and
that can be used in a special team parallel_for.

HostDiagnostics not valid for GPU.
HostDiagnostics is not GPUable but it is derived
from Diagnostics which is.  Just assigning a
HostDiangnostics to a Diagnostics reference
will cause only the GPUable part to be lambda
captured which fixes this error.