Run the rename process test on GPU.
Created a simple process test with a run_() method
for the GPU.  This demonstrated lots of problems
with how the host/device memory was being used and
with the creation and destruction of atmosphere,
diagnostic and other classes. Since these are classes
of Views, care must be taken use to use them properly.
One thing to avoid is the deep_copy() of views of
classes that contain views.  Views of these classes
are useful since a single process has lots of instance
of atmosphere, one for each column, to keep track of.