Create a test with multiple columns to iterate a process over.
This test is an example of a global process, one that iterates
over 1,000 columns to create diagnostic, atmospheric and other
class instances for each column.  Then a Kokkos league is created
with a team for each column and a simple process->run() is call
on each team/column.  This is supposed to reflect a possible
implementation of an atmospheric model running on a single
processor.  Not MPI communication is used.

Since the Haero classes contain Kokkos::Views that we want
to use on device, care must be taken to copy the class
instances to device.  The test loops over the number of columns,
creating class instances and copying to device by lambda capture
and storing the result in a device Kokkos::View as well as a
host std::vector.  The device view can be used in the parallel_for
over all teams in the league.  The host vector can be used to
deep_copy the result back to host when needed, although the
idea is to pass the device instances to many processes without
copying back to host.

The test contains a simple process that demonstrates how
a team instance can be used for loops over levels for the
case that each level is not independent and column processing
is needed, as opposed to calling a simple function that would
act upon each level independently.