Bringing Fortran and C++ modal aero config types closer together.
I'm changing some names of fields and functions within the Fortran
modal_aerosol_config_t derived type to make it easier for us to
understand the correspondence between it and its ModalAerosolConfig
C++ counterpart.

I apologize for changing these things so late in the game, but I do
believe it will help us in the long term.