Duplicate the Skywalker test form MAMRefactor
This is used to show that Haero gets the same numeric
values as the version in MAMRefactor.  This is the first
simple test that does not exercise the code very much
but helps set up the infrastructure. The results were tested
by just diffing with the results from MAMRefactor to demonstrate
the generated python file is identical.