Add first gas/aerosol exchange function gas_aer_uptkrates_1box1gas.
This is the first function called by the gas/aerosol exchange
processes and the first that can be tested against the same
version in the mam refactor repository.  The next step will
be to add Skywalker testing to compare this C++ version in Haero
with the Fortran version in mam_refactor.