Add gas/aerosol exchange test for uptake rates for 1box1gas function.
This adds the unit test and a validation test for the gas/aerosol
exchange process. This first validation test is for the single
function gas_aer_uptkrates_1box1gas and is not an extensive test
but does go through the Skywalker infrastructure to generate
a text file with numeric results that is then compared to a reference
result. This is mostly to lay out the structure of where files
should go, how to compile and link and how to set up a test.