Remove adding optimization options to the compiler flags.
When these flags are added there are two different optimizations
on the compile lines for a Release build, -O2 and -O3.
This causes a compiler warning.

Removing the flags does not eliminate the warning because
ekat also adds flags but now the warning is about the same
flag, -O3, appearing twice which is a step in the right direction.