Strip down AeroSpecies class for Cuda compiler.
The Cuda compiler is very particular about classes
that are allocated as static objects on device.
Such classes can only be initialized with aggregate
initialization and not have non-default constructors
or destructors. So strip this class down to just
data and let the compiler create all the default
functions that might be needed.